refactor cluster
