+++ cluster/calcium/create_container_test.go
@@ -1,6 +1,7 @@
 package calcium
 
 import (
+	"context"
 	"fmt"
 	"testing"
 
@@ -17,7 +18,7 @@ func TestPullImage(t *testing.T) {
 		t.Fatal(err)
 	}
 
-	if err := pullImage(nodes[0], image, ""); err != nil {
+	if err := pullImage(context.Background(), nodes[0], image, ""); err != nil {
 		t.Fatal(err)
 	}
 }
