+++ cluster/calcium/helper.go
@@ -1,72 +1,33 @@
 package calcium
 
 import (
+	"bytes"
 	"encoding/base64"
 	"encoding/json"
 	"fmt"
 	"io"
 	"io/ioutil"
 	"os"
+	"path/filepath"
 	"strings"
+	"text/template"
 
 	"github.com/docker/distribution/reference"
 	enginetypes "github.com/docker/docker/api/types"
 	enginecontainer "github.com/docker/docker/api/types/container"
+	"github.com/docker/docker/api/types/filters"
 	engineapi "github.com/docker/docker/client"
+	"github.com/docker/docker/pkg/archive"
 	"github.com/docker/docker/registry"
 	"github.com/projecteru2/core/lock"
 	"github.com/projecteru2/core/types"
 	"github.com/projecteru2/core/utils"
+	log "github.com/sirupsen/logrus"
 	"golang.org/x/net/context"
 )
 
-func (c *calcium) makeMemoryPriorSetting(memory int64, cpu float64) enginecontainer.Resources {
-	resource := enginecontainer.Resources{}
-	if cpu > 0 {
-		resource.CPUPeriod = utils.CpuPeriodBase
-		resource.CPUQuota = int64(cpu * float64(utils.CpuPeriodBase))
-	}
-	if memory != -1 {
-		resource.Memory = memory
-		resource.MemorySwap = memory
-	}
-	return resource
-}
-
-func (c *calcium) makeCPUPriorSetting(quota types.CPUMap) enginecontainer.Resources {
-	// calculate CPUShares and CPUSet
-	// scheduler won't return more than 1 share quota
-	// so the smallest share is the share numerator
-	shareQuota := c.config.Scheduler.ShareBase
-	cpuids := []string{}
-	for cpuid, share := range quota {
-		cpuids = append(cpuids, cpuid)
-		if share < shareQuota {
-			shareQuota = share
-		}
-	}
-	cpuShares := int64(float64(shareQuota) / float64(c.config.Scheduler.ShareBase) * float64(utils.CpuShareBase))
-	cpuSetCpus := strings.Join(cpuids, ",")
-	resource := enginecontainer.Resources{
-		CPUShares:  cpuShares,
-		CpusetCpus: cpuSetCpus,
-	}
-	return resource
-}
-
-func (c *calcium) calculateCPUUsage(container *types.Container) float64 {
-	var full, fragment int64
-	for _, usage := range container.CPU {
-		if usage == c.config.Scheduler.ShareBase {
-			full++
-			continue
-		}
-		fragment += usage
-	}
-	return float64(full) + float64(fragment)/float64(c.config.Scheduler.ShareBase)
-}
-
-func (c *calcium) Lock(name string, timeout int) (lock.DistributedLock, error) {
+//Lock is lock for calcium
+func (c *Calcium) Lock(name string, timeout int) (lock.DistributedLock, error) {
 	lock, err := c.store.CreateLock(name, timeout)
 	if err != nil {
 		return nil, err
@@ -75,6 +36,89 @@ func (c *calcium) Lock(name string, timeout int) (lock.DistributedLock, error) {
 		return nil, err
 	}
 	return lock, nil
+}
+
+//create container begin
+func makeMemoryPriorSetting(memory int64, cpu float64) enginecontainer.Resources {
+	resource := enginecontainer.Resources{}
+	if cpu > 0 {
+		resource.CPUPeriod = CpuPeriodBase
+		resource.CPUQuota = int64(cpu * float64(CpuPeriodBase))
+	}
+	if memory != -1 {
+		resource.Memory = memory
+		resource.MemorySwap = memory
+	}
+	return resource
+}
+
+func makeCPUPriorSetting(shareBase int64, quota types.CPUMap) enginecontainer.Resources {
+	// calculate CPUShares and CPUSet
+	// scheduler won't return more than 1 share quota
+	// so the smallest share is the share numerator
+	shareQuota := shareBase
+	cpuids := []string{}
+	for cpuid, share := range quota {
+		cpuids = append(cpuids, cpuid)
+		if share < shareQuota {
+			shareQuota = share
+		}
+	}
+	cpuShares := int64(float64(shareQuota) / float64(shareQuota) * float64(CpuShareBase))
+	cpuSetCpus := strings.Join(cpuids, ",")
+	resource := enginecontainer.Resources{
+		CPUShares:  cpuShares,
+		CpusetCpus: cpuSetCpus,
+	}
+	return resource
+}
+
+//relloc container begin
+func calculateCPUUsage(shareBase int64, container *types.Container) float64 {
+	var full, fragment int64
+	for _, usage := range container.CPU {
+		if usage == shareBase {
+			full++
+			continue
+		}
+		fragment += usage
+	}
+	return float64(full) + float64(fragment)/float64(shareBase)
+}
+
+//image begin
+// MakeAuthConfigFromRemote Calculate encoded AuthConfig from registry and eru-core config
+// See https://github.com/docker/cli/blob/16cccc30f95c8163f0749eba5a2e80b807041342/cli/command/registry.go#L67
+func makeEncodedAuthConfigFromRemote(authConfigs map[string]types.AuthConfig, remote string) (string, error) {
+	ref, err := reference.ParseNormalizedNamed(remote)
+	if err != nil {
+		return "", err
+	}
+
+	// Resolve the Repository name from fqn to RepositoryInfo
+	repoInfo, err := registry.ParseRepositoryInfo(ref)
+	if err != nil {
+		return "", err
+	}
+
+	serverAddress := repoInfo.Index.Name
+	if authConfig, exists := authConfigs[serverAddress]; exists {
+		if encodedAuth, err := encodeAuthToBase64(authConfig); err == nil {
+			return encodedAuth, nil
+		}
+		return "", err
+	}
+	return "dummy", nil
+}
+
+// EncodeAuthToBase64 serializes the auth configuration as JSON base64 payload
+// See https://github.com/docker/cli/blob/master/cli/command/registry.go#L41
+func encodeAuthToBase64(authConfig types.AuthConfig) (string, error) {
+	buf, err := json.Marshal(authConfig)
+	if err != nil {
+		return "", err
+	}
+	return base64.URLEncoding.EncodeToString(buf), nil
 }
 
 func makeCPUAndMem(nodes []*types.Node) map[string]types.CPUAndMem {
@@ -128,6 +172,7 @@ func makeMountPaths(opts *types.DeployOptions) ([]string, map[string]struct{}) {
 	}
 
 	// /proc/sys
+	//TODO 外部传入
 	volumes["/writable-proc/sys"] = struct{}{}
 	binds = append(binds, "/proc/sys:/writable-proc/sys:rw")
 	volumes["/writable-sys/kernel/mm/transparent_hugepage"] = struct{}{}
@@ -173,36 +218,188 @@ func execuateInside(client *engineapi.Client, ID, cmd, user string, env []string
 	return b, nil
 }
 
-// MakeAuthConfigFromRemote Calculate encoded AuthConfig from registry and eru-core config
-// See https://github.com/docker/cli/blob/16cccc30f95c8163f0749eba5a2e80b807041342/cli/command/registry.go#L67
-func (c *calcium) MakeEncodedAuthConfigFromRemote(remote string) (string, error) {
-	ref, err := reference.ParseNormalizedNamed(remote)
-	if err != nil {
-		return "", err
+// Pull an image
+func pullImage(ctx context.Context, node *types.Node, image, auth string) error {
+	log.Debugf("[pullImage] Pulling image %s", image)
+	if image == "" {
+		return fmt.Errorf("Goddamn empty image, WTF?")
 	}
 
-	// Resolve the Repository name from fqn to RepositoryInfo
-	repoInfo, err := registry.ParseRepositoryInfo(ref)
+	pullOptions := enginetypes.ImagePullOptions{RegistryAuth: auth}
+	outStream, err := node.Engine.ImagePull(ctx, image, pullOptions)
 	if err != nil {
-		return "", err
+		log.Errorf("[pullImage] Error during pulling image %s: %v", image, err)
+		return err
 	}
-
-	serverAddress := repoInfo.Index.Name
-	if authConfig, exists := c.config.Docker.AuthConfigs[serverAddress]; exists {
-		if encodedAuth, err := EncodeAuthToBase64(authConfig); err == nil {
-			return encodedAuth, nil
-		}
-		return "", err
-	}
-	return "dummy", nil
+	ensureReaderClosed(outStream)
+	log.Debugf("[pullImage] Done pulling image %s", image)
+	return nil
 }
 
-// EncodeAuthToBase64 serializes the auth configuration as JSON base64 payload
-// See https://github.com/docker/cli/blob/master/cli/command/registry.go#L41
-func EncodeAuthToBase64(authConfig types.AuthConfig) (string, error) {
-	buf, err := json.Marshal(authConfig)
+func makeErrorBuildImageMessage(err error) *types.BuildImageMessage {
+	return &types.BuildImageMessage{Error: err.Error()}
+}
+
+func createTarStream(path string) (io.ReadCloser, error) {
+	tarOpts := &archive.TarOptions{
+		ExcludePatterns: []string{},
+		IncludeFiles:    []string{"."},
+		Compression:     archive.Uncompressed,
+		NoLchown:        true,
+	}
+	return archive.TarWithOptions(path, tarOpts)
+}
+
+func makeCommonPart(build *types.Build) (string, error) {
+	tmpl := template.Must(template.New("common").Parse(commonTmpl))
+	out := bytes.Buffer{}
+	if err := tmpl.Execute(&out, build); err != nil {
+		return "", err
+	}
+	return out.String(), nil
+}
+
+func makeUserPart(opts *types.BuildOptions) (string, error) {
+	tmpl := template.Must(template.New("user").Parse(userTmpl))
+	out := bytes.Buffer{}
+	if err := tmpl.Execute(&out, opts); err != nil {
+		return "", err
+	}
+	return out.String(), nil
+}
+
+func makeMainPart(opts *types.BuildOptions, build *types.Build, from string, commands, copys []string) (string, error) {
+	var buildTmpl []string
+	common, err := makeCommonPart(build)
 	if err != nil {
 		return "", err
 	}
-	return base64.URLEncoding.EncodeToString(buf), nil
+	buildTmpl = append(buildTmpl, from, common)
+	if len(copys) > 0 {
+		buildTmpl = append(buildTmpl, copys...)
+	}
+	if len(commands) > 0 {
+		buildTmpl = append(buildTmpl, commands...)
+	}
+	return strings.Join(buildTmpl, "\n"), nil
+}
+
+// Dockerfile
+func createDockerfile(dockerfile, buildDir string) error {
+	f, err := os.Create(filepath.Join(buildDir, "Dockerfile"))
+	if err != nil {
+		return err
+	}
+	defer f.Close()
+	_, err = f.WriteString(dockerfile)
+	return err
+}
+
+// Image tag
+// 格式严格按照 Hub/HubPrefix/appname:version 来
+func createImageTag(config types.DockerConfig, appname, version string) string {
+	prefix := strings.Trim(config.Namespace, "/")
+	if prefix == "" {
+		return fmt.Sprintf("%s/%s:%s", config.Hub, appname, version)
+	}
+	return fmt.Sprintf("%s/%s/%s:%s", config.Hub, prefix, appname, version)
+}
+
+// 只要一个image的前面, tag不要
+func normalizeImage(image string) string {
+	if strings.Contains(image, ":") {
+		t := strings.Split(image, ":")
+		return t[0]
+	}
+	return image
+}
+
+// 清理一个node上的这个image
+// 只清理同名字不同tag的
+// 并且保留最新的 count 个
+func cleanImageOnNode(ctx context.Context, node *types.Node, image string, count int) error {
+	log.Debugf("[cleanImageOnNode] node: %s, image: %s", node.Name, strings.Split(image, ":")[0])
+	imgListFilter := filters.NewArgs()
+	image = normalizeImage(image)
+	imgListFilter.Add("reference", image) // 相同repo的image
+	images, err := node.Engine.ImageList(ctx, enginetypes.ImageListOptions{Filters: imgListFilter})
+	if err != nil {
+		return err
+	}
+
+	if len(images) < count {
+		return nil
+	}
+
+	images = images[count:]
+	log.Debugf("[cleanImageOnNode] Delete Images: %v", images)
+
+	for _, image := range images {
+		_, err := node.Engine.ImageRemove(ctx, image.ID, enginetypes.ImageRemoveOptions{
+			Force:         false,
+			PruneChildren: true,
+		})
+		if err != nil {
+			log.Errorf("[cleanImageOnNode] Node %s ImageRemove error: %s, imageID: %s", node.Name, err, image.ID)
+		}
+	}
+	return nil
+}
+
+func makeCopyMessage(id, status, name, path string, err error, data io.ReadCloser) *types.CopyMessage {
+	return &types.CopyMessage{
+		ID:     id,
+		Status: status,
+		Name:   name,
+		Path:   path,
+		Error:  err,
+		Data:   data,
+	}
+}
+
+func reSetContainer(ID string, node *types.Node, config enginecontainer.UpdateConfig) error {
+	_, err := node.Engine.ContainerUpdate(context.Background(), ID, config)
+	return err
+}
+
+func filterNode(node *types.Node, labels map[string]string) bool {
+	if node.Labels == nil && labels == nil {
+		return true
+	} else if node.Labels == nil && labels != nil {
+		return false
+	} else if node.Labels != nil && labels == nil {
+		return true
+	}
+
+	for k, v := range labels {
+		if d, ok := node.Labels[k]; !ok {
+			return false
+		} else if d != v {
+			return false
+		}
+	}
+	return true
+}
+
+func getNodesInfo(cpuAndMemData map[string]types.CPUAndMem) []types.NodeInfo {
+	result := []types.NodeInfo{}
+	for nodeName, cpuAndMem := range cpuAndMemData {
+		cpuRate := int64(len(cpuAndMem.CpuMap)) * CpuPeriodBase
+		n := types.NodeInfo{
+			CPUAndMem: cpuAndMem,
+			Name:      nodeName,
+			CPURate:   cpuRate,
+			Capacity:  0,
+			Count:     0,
+			Deploy:    0,
+		}
+		result = append(result, n)
+	}
+	return result
+}
+
+func parseStatusKey(key string) (string, string, string, string) {
+	parts := strings.Split(key, "/")
+	l := len(parts)
+	return parts[l-4], parts[l-3], parts[l-2], parts[l-1]
 }
