+++ cluster/calcium/meta_test.go
@@ -15,7 +15,7 @@ import (
 func TestListPods(t *testing.T) {
 	store := &mockstore.MockStore{}
 	config := types.Config{}
-	c := &calcium{store: store, config: config, scheduler: simplescheduler.New(), network: sdn.New(), source: gitlab.New(config)}
+	c := &Calcium{store: store, config: config, scheduler: simplescheduler.New(), network: sdn.New(), source: gitlab.New(config)}
 
 	store.On("GetAllPods").Return([]*types.Pod{
 		&types.Pod{Name: "pod1", Desc: "desc1"},
@@ -36,7 +36,7 @@ func TestListPods(t *testing.T) {
 func TestAddPod(t *testing.T) {
 	store := &mockstore.MockStore{}
 	config := types.Config{}
-	c := &calcium{store: store, config: config, scheduler: simplescheduler.New(), network: sdn.New(), source: gitlab.New(config)}
+	c := &Calcium{store: store, config: config, scheduler: simplescheduler.New(), network: sdn.New(), source: gitlab.New(config)}
 
 	store.On("AddPod", "pod1", "", "desc1").Return(&types.Pod{Name: "pod1", Favor: "MEM", Desc: "desc1"}, nil)
 	store.On("AddPod", "pod2", "", "desc2").Return(nil, fmt.Errorf("Etcd Error"))
@@ -55,7 +55,7 @@ func TestAddPod(t *testing.T) {
 func TestGetPods(t *testing.T) {
 	store := &mockstore.MockStore{}
 	config := types.Config{}
-	c := &calcium{store: store, config: config, scheduler: simplescheduler.New(), network: sdn.New(), source: gitlab.New(config)}
+	c := &Calcium{store: store, config: config, scheduler: simplescheduler.New(), network: sdn.New(), source: gitlab.New(config)}
 
 	store.On("GetPod", "pod1").Return(&types.Pod{Name: "pod1", Desc: "desc1"}, nil).Once()
 	store.On("GetPod", "pod2").Return(nil, fmt.Errorf("Not found")).Once()
